Refactor Spark RDD. Move static methods to separate util class
* Tests are currently failing. Working on this
